added src/components/, encapsulated hint snackbar into a component
