+++ src/constants/Constants.js
@@ -1,14 +1,13 @@
 export default {
   LOCATIONS: {
     DEFAULT: { latitude: 0, longitude: 0 },
-    SINGAPORE: { latitude: 1.3521, longitude: 103.8198 },
   },
   MAP_CAMERA: {
     SINGAPORE: {
-      center: LOCATIONS.SINGAPORE,
+      center: { latitude: 1.3521, longitude: 103.8198 },
       pitch: 0,
       heading: 0,
-      altitude: 0,
+      altitude: 100,
       zoom: 10,
     },
   },
